requirements.txt
# 核心依赖 (仅CPU版本)
torch==2.3.1+cpu                  # PyTorch CPU版 (目前PyTorch官网稳定CPU版本)
--extra-index-url https://download.pytorch.org/whl/cpu

sentence-transformers==3.0.1

# 科学计算基础库
numpy==1.26.4
faiss-cpu==1.12.0
scikit-learn==1.4.2

# 服务与API
fastapi==0.115.0
uvicorn[standard]==0.30.6

# 数据库与搜索
mysql-connector-python==8.3.0
elasticsearch==8.13.0

# 辅助工具
beautifulsoup4==4.12.3
colorama==0.4.6
httpx==0.27.2
lxml==5.3.0
markdownify==0.13.1
overpy==0.7
python-dotenv==1.0.1
requests==2.32.3
rich==13.8.1
tqdm==4.66.2
